+++ index.ts
@@ -80,122 +80,9 @@ const ctx = canvasElement.getContext('2d');
 
 
 
-// Collection of 4 points drawn as a face
-// Try to keep them all on one plane, please!
-
-// I guess, implicitly implements Drawable, Positional VIA extending Quad
-class Square extends Quad {
-  constructor(public centerPoint: Point, public size: number) {
-    let hs = size / 2;
-    let pp = new Point(hs, hs, 0);
-    let pn = new Point(hs, -hs, 0);
-    let np = new Point(-hs, hs, 0);
-    let nn = new Point(-hs, -hs, 0);
-    super(pp, np, nn, pn);
-    this.translate(centerPoint.toVector());
-  }
-}
-
-// Collection of quads
-
 // 6-sided shape with conjoined equilateral faces
 // like, uh, a dice
-class Cube extends Shape {
-  constructor(public centerPosition: Point, public size: number) {
-    super();
-    this.init();
-  }
 
-  init() {
-    let preRot: Angle = null;
-    if (this.quads.length > 0) {
-      this.centerPosition = this.center;
-    }
-    this.quads = [];
-    let hs = this.size / 2;
-    /* 
-    // xyz, p = positive, n = negative
-    let ppp = new Point(hs, hs, hs);
-    let ppn = new Point(hs, hs, -hs);
-    let pnp = new Point(hs, -hs, hs);
-    let pnn = new Point(hs, -hs, -hs);
-    let npp = new Point(-hs, hs, hs);
-    let npn = new Point(-hs, hs, -hs);
-    let nnp = new Point(-hs, -hs, hs);
-    let nnn = new Point(-hs, -hs, -hs);
-    // Front
-    this.addQuad(new Quad(ppn, npn, nnn, pnn));
-    // Back
-    this.addQuad(new Quad(ppp, pnp, nnp, npp));
-    // Left
-    this.addQuad(new Quad(npn, npp, nnp, nnn));
-    // Right
-    this.addQuad(new Quad(ppn, pnn, pnp, ppp));
-    // Top
-    this.addQuad(new Quad(ppp, npp, npn, ppn));
-    // Bottom
-    this.addQuad(new Quad(pnp, pnn, nnn, nnp));
-    */
-    // Old, easy system caused quads to share point references, which messed up certain translations,
-    // eg, made it rotate ~3x faster because every point was rotated 3 times.
-    // Front
-    this.addQuad(
-      new Quad(
-        new Point(hs, hs, -hs),
-        new Point(-hs, hs, -hs),
-        new Point(-hs, -hs, -hs),
-        new Point(hs, -hs, -hs)
-      )
-    );
-    // Back
-    this.addQuad(
-      new Quad(
-        new Point(hs, hs, hs),
-        new Point(hs, -hs, hs),
-        new Point(-hs, -hs, hs),
-        new Point(-hs, hs, hs)
-      )
-    );
-    // Left
-    this.addQuad(
-      new Quad(
-        new Point(-hs, hs, -hs),
-        new Point(-hs, hs, hs),
-        new Point(-hs, -hs, hs),
-        new Point(-hs, -hs, -hs)
-      )
-    );
-    // Right
-    this.addQuad(
-      new Quad(
-        new Point(hs, hs, -hs),
-        new Point(hs, -hs, -hs),
-        new Point(hs, -hs, hs),
-        new Point(hs, hs, hs)
-      )
-    );
-    // Top
-    this.addQuad(
-      new Quad(
-        new Point(hs, hs, hs),
-        new Point(-hs, hs, hs),
-        new Point(-hs, hs, -hs),
-        new Point(hs, hs, -hs)
-      )
-    );
-    // Bottom
-    this.addQuad(
-      new Quad(
-        new Point(hs, -hs, hs),
-        new Point(hs, -hs, -hs),
-        new Point(-hs, -hs, -hs),
-        new Point(-hs, -hs, hs)
-      )
-    );
-
-    this.translate(this.centerPosition.toVector());
-  }
-}
 
 // Test driving code //
 let canvas = new Canvas(canvasElement, Angle.fromDegrees(45));
